+++ source/rtcw_mp/client/cl_parse.cpp
@@ -390,7 +390,7 @@ void CL_ParseSnapshot( QMsg *msg ) {
 	// read areamask
 	len = msg->ReadByte();
 
-	if ( len > sizeof( newSnap.areamask ) ) {
+	if ( len > (int)sizeof( newSnap.areamask ) ) {
 		Com_Error( ERR_DROP,"CL_ParseSnapshot: Invalid size %d for areamask.", len );
 		return;
 	}
@@ -634,7 +634,7 @@ void CL_ParseDownload( QMsg *msg ) {
 	}
 
 	size = msg->ReadShort();
-	if ( size < 0 || size > sizeof( data ) ) {
+	if ( size < 0 || size > (int)sizeof( data ) ) {
 		Com_Error( ERR_DROP, "CL_ParseDownload: Invalid size %d for download chunk.", size );
 		return;
 	}
