+++ source/rtcw_mp/server/sv_main.cpp
@@ -99,7 +99,7 @@ char    *SV_ExpandNewlines( char *in ) {
 	int l;
 
 	l = 0;
-	while ( *in && l < sizeof( string ) - 3 ) {
+	while ( *in && l < (int)sizeof( string ) - 3 ) {
 		if ( *in == '\n' ) {
 			string[l++] = '\\';
 			string[l++] = 'n';
@@ -409,7 +409,7 @@ void SVC_Status( netadr_t from ) {
 			String::Sprintf( player, sizeof( player ), "%i %i \"%s\"\n",
 						 ps->persistant[WMPERS_SCORE], cl->ping, cl->name );
 			playerLength = String::Length( player );
-			if ( statusLength + playerLength >= sizeof( status ) ) {
+			if ( statusLength + playerLength >= (int)sizeof( status ) ) {
 				break;      // can't hold any more
 			}
 			String::Cpy( status + statusLength, player );
@@ -468,7 +468,7 @@ void SVC_GameCompleteStatus( netadr_t from ) {
 			String::Sprintf( player, sizeof( player ), "%i %i \"%s\"\n",
 						 ps->persistant[WMPERS_SCORE], cl->ping, cl->name );
 			playerLength = String::Length( player );
-			if ( statusLength + playerLength >= sizeof( status ) ) {
+			if ( statusLength + playerLength >= (int)sizeof( status ) ) {
 				break;      // can't hold any more
 			}
 			String::Cpy( status + statusLength, player );
