+++ source/rtcw_sp/qcommon/common.cpp
@@ -1146,7 +1146,7 @@ void Hunk_FreeTempMemory( void *buf ) {
 
 
 	hdr = ( (hunkHeader_t *)buf ) - 1;
-	if ( hdr->magic != HUNK_MAGIC ) {
+	if ( hdr->magic != (int)HUNK_MAGIC ) {
 		Com_Error( ERR_FATAL, "Hunk_FreeTempMemory: bad magic" );
 	}
